requirements.txt
amqp==5.0.9
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-cors-headers==3.10.1
django-debug-toolbar==3.2.2
django-filter==21.1
django-redis==5.2.0
django-silk==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
dnspython==2.1.0
docopt==0.6.2
drf-nested-routers==0.93.4
eventlet==0.33.0
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flower==1.0.0
gevent==21.12.0
geventhttpclient==1.5.3
gprof2dot==2021.2.21
greenlet==1.1.2
gunicorn==20.1.0
humanize==3.13.1
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.3
locust==2.5.1
MarkupSafe==2.0.1
model-bakery==1.4.0
msgpack==1.0.3
mysqlclient==2.0.3
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
prometheus-client==0.12.0
prompt-toolkit==3.0.24
psutil==5.9.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pywin32==227
pyzmq==22.3.0
redis==4.1.0
requests==2.26.0
requests-oauthlib==1.3.0
roundrobin==0.0.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
watchdog==2.1.6
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
zope.event==4.5.0
zope.interface==5.4.0
